feat(command): add anonymous argument
Add non-named argument that can be anything and not requires a specific
list.
Before, it needed a trick to allows developer to create an argument that
is not a specific list and not an argument that will serve into the
executor (like a coordinate or a name for a thing).

This feature needed to re-write the base of what an argument is.

Also, the way to create an argument has been re-work to have an easier
to create an argument. Now it proceeds step by step :
1. The type of argument
2. The type of sender
3. What argument(s) the argument will have
4. Customize it with custom tab completer type or some preventions.

To finish, the CommandData, now has a immutable list for the list of
string that represents all the arguments (and not the remainings
arguments that is a linked list).
The transformation from a linked list to an immutable list was done to
prevent the developper to perform modification on it (the arguments can
never change once the sender has sent the command).