feat(color): possibility to add custom color code
When adding a color code that can be a Minecraft color code (&x) or a
RGB color code (#xxxxxx), the custom color code will be present
everywhere the colors are used (item, menu, message).

This allows easier development (especially when using RGB code that can
be annonying to type everytime or to remember).
But this is useful to allows dynamic configuration with color code.