fix(paper-color): error when end with color
When a string to color with the paper, if the string ends with a color
code (like #xxxxx or &x), there was an exception.
This was due to the use of an offset when looping through the color and
the text arrays to create component. But if there is no text associated
with the color, the module tries to access to an non-existing text.